+++ packages/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @harbor-design/web
 
+## 0.0.79
+
+### Patch Changes
+
+- Updated dependencies
+  - @harbor-design/proform@1.4.4
+
 ## 0.0.78
 
 ### Patch Changes